Correct val_function SQL to ensure calling the class->method
In generating the encoded string from a functional sample array,
the callback's options were incorrectly identified as needing to
be an array with one of the values to be assigned as a key.

This presents the data as an array of elements and can be properly
processed when broken down by the code.