Move incorporation of language definitions into class constructor.
In order for an error message to be presented, a language define
must exist at the point of processing the submitted request.
To fully support code flexibility, the associated language must
be available which means that inclusion of the language file
must happen after enough information has been gathered to identify
the language that should be used.

The original placement at the top of the file caused execution
when the file was first loaded which may be too soon in the process
depending on the initial load point of including the code.

By placing the language load within the __construct method then
the definitions are delayed and loaded at the point where the
class may be used.