Add boolean configuration checker.
The FILTER_VALIDATE_BOOLEAN by default returns true only if the
value being reviewed is a value that evaluates to a boolean
true. The way this configuration_checker is written, if the
filter_var function returns false, then the error message would
be displayed.

In this modification, if the value evaluates to either true
or false then it is storable without an error, but if it is
neither, then the error message would be displayed.