Perform products_id validate on first load/session not started.
A portion of includes/init_includes/init_sanitize.php is not
executed on the first load to a site, leaving follow on code to
respond instead of this change.  Not sure if this is to in some
way improve first page load time or if wasn't something simply
expected to be handled later but was offerred an opportunity
during routine navigation.