Provide the option of not going to 404 when product status is 0
Split out the code for handling 404 Header responses so it can be intercepted with an Observer if needed.
Added ability to define a constant PRODUCT_THROWS_200_WHEN_DISABLED to boolean `true` without needing an Observer, if a Store wishes to allow a product to display even when disabled.
A 404 will still trigger if the product doesn't actually exist in the db.
A 410 can be triggered by setting `products_status` to `-1` in custom code, without needing an observer for the 410 response.

This change should also reduce queries on product page.

Fixes #1699