Sanitize IP addresses vai filter_var
As identified in [this](https://www.zen-cart.com/showthread.php?225679-PHP-Fatal-error-1406-Data-too-long-for-column-ip_address-at-row-1&p=1359685#post1359685) Zen Cart forum thread.

Use the built-in PHP `filter_var` function to fully vet the incoming IP address.  The change specifies the IP filter types to make the processing more obvious.

If an invalid IP is found, set (as current) the IP value to a single dot and issue a notification allowing follow-on code to catch and possibly record the invalid access.