Incorporate configuration update for email val_function values
As a "last note", wanted to include the RegEx used to locate/
create these install sql statements through a search and replace
method.

Search:
(.*?)(INTO )(.*) \((.*)(val_function)(.*?)VALUES \((.*?\'\, )(.*\')\,(.*\')\,(.*)(\'\{\"error.*\}\').*

Replace:
UPDATE $3 SET $5 = $11 WHERE configuration_key = $8;

This worked for the val_function key, but would need to be
modified for most any other key. There may also be an
easier set of RegEx, but it was something that worked and made
sense to me as I built it. :)