Refactor product_info main_template_vars
Used the new variable `$product_not_found` instead of checking for `isset` related to `$product_info`.  I did ** *not* ** create a new variable (`$product_found`) which could be assigned in line 22 as: `$product_found != $product_not_found = $product_info->EOF;`, although a double negative (`!$product_not_found`) is used in the new code.

Set default values for variables populated when a product is in the database. This is to eliminate individual checks on each variable proposed in the original modification.
This way if a `products_id` is attempted that is not in the database, strict mode will not report notices and likely future PHP versions will not provide associated errors.

Included is a "hint" to potentially use the default setting offered when creating a new product; however, have incorporated the default value(s) of the Zen Cart installation for each associated variable.

Also included is removal of single quotes around items in the query that are expected to be treated as integers to reduce the conversion performed by the database and to match the data type.

These changes may allow removal of some of the other changes made in functions files. Will look at those and merge as necessary.