Select default email recipient
In #2332, an incorrect resolution for strict PHP notices was provided by assigning the default value for dropdown selection to be an `integer`.  At this point in the code, the `customer` is the id associated with individual to be selected when the gv_mail page has een entered. The id is either the e-mail address or the standard name such as `All%20Customers`.  This restores the default selection of email address dropdown.