Exclude "hidden" order-status updates from account-history
While a "hidden" admin-only order-status comment is excluded from display catalog-side, the "Order Status" name was being displayed, which might not be appropriate for customer to see.
This PR changes it to show the last non-hidden order-status-name in the My Account page and the My Account History page (which is accessed by clicking the Show All Orders) from the My Account page).