Remove hard-coded image dimensions
This code controls the size of images in a database-defined column.  If the columns are set to 3, then 33% becomes the column width.
However, with wider screens and the diverse size of mobile/tablet devices, restricting the height and width really only restricts the ability to "fill" any screen with the product being presented.  Removing these two calls return control to the store owner.
I realize this may have been an attempt to combat the use of images that were different in size but, I think that should be scrapped in order to better control display of products images in any environment.