Correct misplaced form closure when only 1 address-book entry
If a customer has only one entry in their address-book, the 'Select current address' section of the `checkout_payment_address` and `checkout_shipping_address` pages renders the `</form>` for that section, even though the `<form>` has not been opened.

The issue is present in both the `responsive_classic` and `template_default` templates' versions.