Eliminate php notice
If the query returns no results, then attempting to obtain `configuration_value` from the query results causes a PHP notice that will become an error in the future.  By assigning a value to `$lookup_value` first and then reassigning to an alternate value if there is a record returned, no notice/error is generated.