Address php notice while awaiting upgrade
Until the `configuration_key` is inserted into the database, a php notice that is expected to become an error in the future is issued.  This at least allows continued execution while the `configuration_key` is still not yet inserted without throwing the notice.  It also will mean that the internal code will not be executed if the `configuration_key` is removed from the database in the future.  I have not yet seen how this `configuration_key` is used in the code, to possibly advise of anything different.