Ensure GET parameters do not begin with amp;
Finding that there are cases where $_GET keys begin with 'amp;'
and that as a result testing against the key without the
prefix of 'amp;' basically prevents finding it... In ZC 1.5.6
and 1.5.7 this appears to cause a product to be considered
non-existent when it is in the database, active, and in stock.

For example: index.php?main_page=product_info&amp;cPath=1_4&amp;products_id=1

causes the store to present the category breadcrumb but then
the product is identified as not present.