Eliminate php notices when viewing gv_faq
Ensure that $_GET['faq_item'] is always set to something to support
the switch/case operation.

Assign default values to variables before they are used in the
template.
Use/apply the review of an individual being logged in or possibly
a guest to at least eliminate notices of testing an unassigned variable.
Either place a value of zero or the value of the $_GET['gv_no'].