Add control if admin switch present and enabled
Allow the enablement of product that have a historical product_available_date if the constant is defined and set to support enablement of this feature.

Consideration was given for testing/checking against `!= '0'`; however, seems that such an option would only be necessary to support some sort of third, fourth, etc... condition which is not clear what it may be and therefore likely would require modification of this area of the code if generated.  By the check specifically against `== '1'`, whatever defines it (file or database) if it is equal to 1 then this area of code will execute.

Alternatively, for those that do mass installations, if they choose not to include the option in the database or choose to remove it, then code will continue to operate without execution of this feature.