Add ability to translate a string provided constant to session's language
Was identified in #3033 that by moving the loading of `$messageStack` to before the `$_SESSION` has been started and before the language has been identified, that messages pushed to `$messageStack` would not be language dependent.  This adds the ability to push the expected language constant as a string that would be translated to the current session language when the session was available and the additional message(s) were then presented.