Bug fix: wrong value_id value
As reported in #2973 This fixes a bug in the options_values_manager.php

This happen on creating a new value when selecting an option from the filter dropdown that contains no values first.

The problem is that the ```$value_id``` sits inside a ```foreach``` that is never fired on empty options. 
Moving the query that sets ```$value_id``` outside the ```foreach``` fixed the error on pressing insert