Respect line if field exists and table is bing modified
Fixes #3278.

While not every possible combination was researched, the SQL statement provided in #3278 is very specific and results in executing the default portion of the function evaluation.  An earlier consideration was to swap the sequence of the `ADD` and `DROP` such that if `DROP` occurred, then the result was the reverse of the field existing in the table so that the function resolved to allow the line to be executed.  This though directly resolves the condition that led to finding the problem...