Streamlining css buttons
Removing unused button code in html_output.php and updating the Classic and Responsive Classic stylesheet_css_buttons.css. This "streamlining" chose to use the cssButton class as default and removed the duplication of the button class. The remaining css in the classic template is as cleaned up and as efficient as possible for css buttons.