Skip loop through products if db query is empty.
`shopping_cart::get_products()`:
When the db query is customized to check only for products which match certain criteria, not aborting this loop will cause incorrect empty entries in the resulting array, and will trigger numerous "index not found" notices.
Checking for validity allows skipping the unnecessary processing of this large section of code and avoids false entries.