Allow text to contain NULL
While reviewing changes being made to this file for 1.5.7, noticed
that there are several fields that use the type 'string' when
using the `$db->bindVars` method.  As a result if the associated
sanitized string contained `"NULL"` in the text (of any length),
the string would be sanitized to a null value for database insertion.
In a standard install this will cause a mysql error as the field(s)
can not be set to `null`.

While I have not specifically used `"NULL"` in any such fields under
normal use, I did validate that such usage would be a problem.

I have no personal need for this revision, but seems like it is
unnecessarily stringent and not otherwise described as a "caveat".