Remove mysql need to convert string to integer.
Where values in a SQL statement are expected to be an integer,
the value is cast to an integer, and was previously surrounded
by single quotes, the additional quotes have been removed and
in a few cases, the value is now cast to `int` and in a few other
cases the `zen_db_input` function sanitization was replaced with
casting to `int`.

There may still be some efficiencies to be gained within a function
to where the original value would be cast to an integer one time
and then used throughout instead of cast to int on each access.

I did not fully touch conversion of SQL statements to use all
capital letters for words such as `SELECT`, `UPDATE`, `SET`, etc...