Restore footer position to original.
No reason has been provided for moving the footer to within
the divider.  Without reason it should not be moved as may/will
cause additional work for coders that affect this page through
use of the footer's notifier.

If there is a reason, then it should be made known public for
incorporation/consideration of use.