Resequence placement of parameter separator
While reviewing code associated with the changes to this file,
noticed that there was at least one instance of a url being
generated to where it would/may end with a single `&`. This
places the parameter separator (`&`) with the parameter that
is being added instead of as an intermediate character or as
a final ending.  Such use is not necessary for the first
parameter as functions such as `zen_href_link` will incorporate
the appropriate first separator as necessary.