Use the built-in ZC function for language icon.
Although the original code of this function used the language's
`name` for the alternate text and `zen_get_language_icon` uses
the language's `directory`, substituted the associated function
instead of effectively duplicating code.  If it is desired to
change the alternate text to be the language's name, then would
suggest doing so in the associated function rather than at
individual code points throughout the software or create an
alternate function/parameter that target's the data to be desired.