Remove conversion from integer cast variables
This removes the excess quotes (string casting) from around
variables that are cast to integers and expected to be stored
in the database as integers.  Previous result was value
(potentially a string) was cast to an integer, then encapsulated
into quotes which makes database treat as a string that then
is to be stored into an integer type datatype so requiring
additional conversion by the database to int. This reduces
the additional two conversions though only one was really
processing related (conversion of string back to integer).