Default/Restore date to NULL and ignore if is NULL
In preparation for future updates to the date/time processing
to use NULL instead of the historical default of `'0001-01-01 00:00:00'`
choosing to use NULL instead which is supported by the field.

This reduces one change to the overall system.  Check for whether
to process the record still compares against `'0001-01-01'`
in the event other "systems" clear the field to be appear as
empty.