Update the installation and upgrade scripts for Automatic and Manual
For new installation (not using a previous database), set this to
automatic to enable product that are disabled and have a future
date.  For upgrades left this as Manual so that a database update
will not automatically enable product that were disabled but had
a future date.

Previous operation without this additional feature would have
left the product disabled from viewing on the store until the
product were enabled.