Speed up the Admin orders listing (remove query join)
Remove join on orders_total table for preparing the paginated list of orders.
The join is an expensive operation in this case because of the way the schema is built.

This change relies on the `order_total` field in the `orders` table to show the amount instead.
The only downside to this change is that any mods doing bespoke changes may cause this data to be out of sync if they fail to update the `orders` table when updating `orders_total` entries. They should have already been doing that anyway, but this will expose their deficiency if it exists.