Add count_key for splitpageresults
There are some queries that require providing a specific field or set of fields when splitting the page results.  While this edit does not add all available parameters of the `splitPageResults` class, it does add the next one.  At this time, the parameters to the class are: `$query, $max_rows, $count_key = '*', $page_holder = 'page', $debug = false, $countQuery = ""` with the first two already considered in the previous code.