Fix Color Contrast in Responsive_Classic
The colors for the present responsive_classic stylesheets_colors.css do not meet the standards for readability in accessibility tests.
An example of where it stood can be seen by going to //webaim.org/resources/contrastchecker/.  Enter #FFFFFF on the right and #05A5CB on the left.  The contrast ratio is 2.89:1 which fails all five situations.  Changing #05A5CB to #364fb5 results in a 7.13:1 ratio and passes all situations in which the font might be used.
Also, changing the hover color for the navCatTabs menu is required to up the contrast during a hover over that menu.
There are other colors affected as well.  One would think that red on white would not be a problem.   Yet, replace the #364fb5 with #FF0000 (red), and the ratio is 3.99:1.  Another fail.  Thus, the standard for specials and special pricing actually makes it difficult for some folks to read the text/price.
This PR is not meant to "cure" anything but the color contrast and therefore only affects includes/templates/responsive_classic/css/stylesheet_colors.css.
It is meant to provide anyone wishing to use the template to have no color contrast hits in accessibility and SEO.