This orphans the `products_viewed` field from the `products_description` table, by relocating the counting to the new `count_product_views` table, where views are tracked by date.
Also skips tracking hits by Admin and Spider visitors.

This removes the "reset" in Store Manager because the former need for a reset is deprecated because data can now be filtered by date.

The report's default date range is "today for the past 30 days". Editable on-screen to filter to a different date range.

NOTE: Plugins which have cloned the `copy_product_confirm.php` file for custom product-types will want to skip the copy of the `products_viewed` field when copying.