Base Font Size Too Small for Accessibility
While there is no specific written standard for font-size, it is considered to be 12pt or 16px.  At the original 72.5% and the fact that most alerts and small text are set to em (some as low as .5em), readability is taking a big hit on the mobiles and things like the site legal info.
Making the setting 80% at least gets us to 10pt and triggers no errors for font too small.