Update for loading javascript and css in admin
This update adds the ability to load multiple javascript and/or css files. The have to start with the page name you want the file to load on.
For example ```admin/includes/javascript/product_extraJscript.js``` will now load on product.php.
The modification will make it easier for mod makers to include their css or javascript to a page.
Also added the ability to load php javascript files