Update Square module to v1.3  - uses SquareConnect SDK version 3.20200528.1
Note: This is the last release of SquareConnect SDK that is compatible with PHP 5. Future versions will require PHP 7.1+, and will be integrated to the SquarePHP SDK (not SquareConnect).
This version is being integrated here into ZC v1.5.7 because v1.5.7 still aims for PHP 5.6 compatibility.  (v1.5.8 will not).

A version 2 of this module will be released to work with the new SDK (and PHP 7.1+ only) in the near future.