This introduces a new naming convention for Observer class method names which respond to Notifier calls.
This allows IDEs (text editors) to run a (case-insensitive) search for a notifier name and find these methods which relate to those notifiers.

The new pattern/flow is that when a Notifier hook is triggered it will fire an Observer method using the first method name matching the following pattern/order:

1. `snake_cased_name_of_notifier_event()` ... as long as the name starts with "notify_" or "notifier_" (as is the standard convention already)
2. `updateCamelCasedNameOfNotifierEvent()`
3. `update()`

In all cases the list of parameters accepted is the same.