Allow observer to be executed from multiple files
While I can not seem to find this as an equivalent operation
for ZC 1.5.8 (too rushed?), I discovered this bug where
if a notifier was observed by more than one observer, then it
seems that either a secondary observer was not executing or
if it was, data was lost... This allows the notifier to be
executed through multiple observers and not lose the data that
was originally transferred.

Issue was seen with installation of both stock by attributes
and products with attributes image swap.  Both were listening
to the notifier: `NOTIFY_ATTRIBUTES_MODULE_START_OPTION` but
only SBA's version was execute while attribute image swap was
not executing and in strict mode there were notices thrown that
variables were not defined/declared...

I'll continue to try to find the ZC 1.5.8 version of this
code section, but not sure if it will/does suffer the same
issue as it has been moved into using namespace type execution.