define TOPMOST_CATEGORY_PARENT_ID = 0
Instead of using `'0'` to denote the "top"/root/parent category, use the `TOPMOST_CATEGORY_PARENT_ID` constant.

This will improve clarity in code when checking whether a `category_id == '0'` vs being 0 for some other reason.
It also allows, in theory, for a different parent/root to be set for other reasons (such as an MPTT category structure, or nested sub-groupings .... neither of which are planned/forecasted on any roadmaps).