Refactor `else if` logic test
Simplified the logic of whether to perform the code within the `else if` section or not.  Basically, if there is an `elseifFunc` then evaluate to determine the result followed by whether to inverse that result or not (inverses result if `elseifNot` is `truish`). If `elseifFunc` is evaluates as `false` when processed through `empty` then the internal code is skipped.