Force conversion to float
Using MySQL 5.7, the cast to float was not being applied to the result of the if statement, but instead basically to the if test and a fatal error to be logged (and possibly/eventually a duplicate order number to be generated). This corrects the casting to apply to whatever result is to be used for `settle_amount`.