Support dropdown default selection
Adds an internal variable to retain the value of $selected_attribute
throughout the loop process for use outside the loop. When the loop is
complete $selected_attribute will be equal to this internally tracked
variable so that the dropdown (and any other external tracking) will
see what the end result is/was for identifying the selected attribute.

Fixes #3734