Add an asterisk to fields
I noticed that in zc 1.5.8 on the admin side there is no longer an asterisk shown on required fields.
These changes add back the asterisk on fields that are set as required through the input or select function.

The asterisk is placed in an input addon span. This ensures the alignment on the right is always the same as fields that are not required