Put admin search box in seperate file
There are multiple pages in the admin that use a search box. The code for these boxes have to be maintained separately on each page, which is a pain when you want to change something.

This update puts the search box in a separate file (```admin\includes\modules\search_box.php```). This file contains mainly html code to display the box.  To display on a page do something like this:
```php
<div class="col-sm-offset-8 col-sm-4">```
  <?php include DIR_WS_MODULES . 'search_box.php'; ?>
</div>
```