Modify the default selected options names values selection for reinsert/update
This is what I requested in #2950.  Operation a time ago was to simply
display the option names values selector with what appeared to be the
first option in the list.  It was recently modified so that the last
option names value listed item on the current page to be the current
selection in the dropdown.  This seems to make sense at least when
choosing to add the first item that is to be listed, but once an option
names value has been added, it would be sensible if the selector remained
the same to support entry of the next value for that option name.

This modifies operation to keep the same name selection from one entry
to the next and also when updating an item.