Prevent page parameter when link goes to page=1
Prevents display/inclusion of `page=1` in the generated web address(es).
This is to remove the added page parameter into the page address as a
result of this being eventually fed to the canonical link.

This change also addresses the parameter and its appearance in the
canonical link if manually added to the web address.  This does a full
removal from the canonical address when page=1 or if the parameter
is non-numeric.

Fixes #4008