Prevents php notice using PluginManager and a plugin has an upgrade available.
I haven't been able to fully figure out why this is happening or rather
why the __construct has `tableDefinition['configBox']` which doesn't seem
to be used and that these methods don't reference that particular key, but
when I tried to substitute the `configBox` value(s) in these modifications
then the pluginmanager would not move forward in providing information
or allowing the next action. Therefore, and empty array seemed to be the only
fix to support moving to the next step.