Support admin use of .svg image extension
I did not identify if a suggestion such as this had previously been denied. This adds the allowance of files with the extension `svg` to be loaded/displayed in the admin area. the `svg` extension was already recognized in the optional caching improvements section.  This also adds expiration control for the same `svg` file extension

This has been suggested in response to Zen Cart forum thread: https://www.zen-cart.com/showthread.php?227351-Use-logo-svg-in-place-of-logo-gif-or-logo-png