Removes inspected session from browser when session no longer exists.
If tracking a particular session, when the session disappears or is removed
it remains to be tracked; however, no longer exists. Further, if the session
being tracked is incorrectly entered, it will appear ass if the session
does not have a cart even if there is one.

This becomes a visible problem if there are two or more sessions active,
one of them is being tracked and the page is set to refresh periodically.
When that session expires and eventually only one session exists, at that
point, it is not possible to click on a session to force tracking it.

This clears the tracking data if it doesn't exist.