Refactor ot_coupon to expose validation methods
This refactoring segments several big long blocks of code into smaller functions.

The end result is that it now allows for some external processes to use the module's coupon validation rules.

For example, it would be fairly easy to capture a ?coupon=FOO from the URL and ask ot_coupon whether it is valid, and if it's valid then apply it to the Session (eg, cart) automatically.
Handy for emailings or in-site clickable links.

Backward compatible with v1.5.7 save for a small change in the data passed to the NOTIFY_OT_COUPON_CALCS_FINISHED notifier that was added in 1.5.7.