Eliminate $attributes notice on actionUpdateProduct decrease
Address notice at line 1789, ensuring that $attributes is set to an array of some type, either empty or including attribute data. Formerly this looked for the "presence" of the attribute; however, both in current PHP environments and when strict PHP is enabled a notice was expected to be generated when entering this code cycle.