Update shopping_cart.php
Corrects potential notice associated with product being evaluated if it is not within the `$_POST`ed content (for whatever unusual reason that may exist).
Corrects notice when evaluating decrease but the quantity increases or remains the same or in reverse if evaluating for increase and the quantity decreases or remains the same.