Prevents notice during module setting.
Reproducible by accessing the modules page (e.g. shipping) from some other
page in the admin, editing the default selected module and saving that
module without navigating to anything else.  During this process, the
uri parameter of module will not be set and later cause a notice on line 65.

Had previously submitted a fix for line 65; however, recently was able
to recreate the issue directly and specifically.