Parameter for zen_redirect of zen_href_link incorrect.
The admin version of `zen_redirect` has only one parameter which is the
destination address. The generation of that address through `zen_href_link`
accepts an indicator of whether to use `SSL` or `NONSSL`. That value is
expected to be in the third position of calling the `zen_href_link`
function. The modified line originally has `NONSSL` as the second parameter
to the `zen_redirect` function. When modified as provided, this would
correctly become the third parameter.