Fix "Uncaught TypeError: mysqli_close(): Argument #1 ($mysql) must be of type mysqli, bool given"
When invalid/missing db credentials are used to access the database, it is possible for this error to be triggered:
PHP Fatal error:  Uncaught TypeError: mysqli_close(): Argument #1 ($mysql) must be of type mysqli, bool given in /includes/classes/db/mysql/query_factory.php:603

This is most likely to occur on new installs where no configuration yet exists.

Prior to PHP 8 this would fail silently. In PHP 8 it is treated as a fatal error, presumably because it falls into "Uncaught exceptions now go through "clean shutdown", which means that destructors will be called after an uncaught exception."