Remove php notice when searching for non-existent description
If function is used to look up tax rates based on a description, even
TEXT_UNKNOWN_TAX_RATE is not present and therefore leads to the problem
described at this ZC forum thread:
https://www.zen-cart.com/showthread.php?228051-tax_rate-undefined-index

Chose the specific use of setting the tax rate to 0.0 so that one could
"easily" consider other rates to apply if necessary or some other
variation. Another alternative to this could have been to upfront
evaluate `$description` and if it equaled TEXT_UNKNOWN_TAX_RATE to then
continue the loop which would skip the tax assignment.