Support product page return if file upload problem
If there is a problem with the upload of a file as determined by
the upload class' parse method, then regardless the page to which to
return this presents a message and stops the addition of the remainder of
the product conditions.

Without this code, the product could be added to the cart without the
uploaded information (specifically removed because of issue bringing the
code to this section) even if the file upload attribute was marked as
required.

Discussed in Zen Cart forum at: https://www.zen-cart.com/showthread.php?228049-items-added-to-Shopping-Cart-despite-errors