Permit storing an empty email address
Allows clearing the email address information. Previous implementation would
identify an empty email address as an error and prevent removing the address
which basically meant that once an email address was entered, the only
other recourse was to provide some email address that "looked" right in
order for emails *not* to be sent to that address.