application_bootstrap.php - loading of plugin functions
I am not sure that this is the correct way to deal with this issue and whether other lines need to be added as well.
i.e. more than just functions ??
I am not aware of the overall thinking here so am just really pointing out a possible area of error.